MS4 - Many files of project - Add docstring
