+++ backend/src/Controller/Api/Private/PointController.php
@@ -18,11 +18,26 @@ class PointController extends AbstractController
 {
     private $security;
 
+    /**
+     * Constructs a new instance of the class and initializes the security property.
+     *
+     * @param Security $security The security service to be used.
+     */
     public function __construct(Security $security)
     {
         $this->security = $security;
     }
 
+    /**
+     * Adds a new point to the database.
+     *
+     * @param Request                $request The HTTP request.
+     * @param EntityManagerInterface $entityManager The entity manager.
+     * @return JsonResponse The JSON response.
+     * @throws \Doctrine\DBAL\Exception\ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/points', name: 'apiAddPoint', methods: ['POST'])]
     public function apiPointAdd(
         Request                $request,
@@ -74,6 +89,14 @@ class PointController extends AbstractController
         }
     }
 
+    /**
+     * Retrieves all points for the authenticated user and returns them in JSON format.
+     *
+     * @return JsonResponse The JSON response containing the points data.
+     * @throws \Doctrine\DBAL\Exception\ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/points', name: 'apiGetPoints', methods: ['GET'])]
     public function apiGetAllPoints(): JsonResponse
     {
@@ -110,6 +133,15 @@ class PointController extends AbstractController
         }
     }
 
+    /**
+     * Retrieves a point by its ID and returns it as a JSON response.
+     *
+     * @param Points $point The point object to retrieve.
+     * @return JsonResponse The JSON response containing the point data.
+     * @throws ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/points/{id}', name: 'apiGetPoint', methods: ['GET'])]
     public function apiGetPointById(Points $point): JsonResponse
     {
@@ -137,6 +169,17 @@ class PointController extends AbstractController
         }
     }
 
+    /**
+     * Updates a point in the database.
+     *
+     * @param Request $request The HTTP request object.
+     * @param Points $point The point to be updated.
+     * @param EntityManagerInterface $entityManager The entity manager.
+     * @throws ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     * @return JsonResponse The JSON response containing the updated point data or an error message.
+     */
     #[Route('/points/{id}', name: 'api_points_update', methods: ['PUT'])]
     public function update(
         Request                $request,
@@ -191,6 +234,16 @@ class PointController extends AbstractController
         }
     }
 
+    /**
+     * Deletes a point by its ID and returns a JSON response indicating the status of the deletion.
+     *
+     * @param Points $point The point object to delete.
+     * @param EntityManagerInterface $entityManager The entity manager used to delete the point.
+     * @return JsonResponse The JSON response indicating the status of the deletion.
+     * @throws ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/points/{id}', name: 'api_points_delete', methods: ['DELETE'])]
     public function delete(
         Points                 $point,
