+++ backend/src/Controller/Api/Private/PointTelemetryController.php
@@ -20,6 +20,15 @@ class PointTelemetryController extends AbstractController
         $this->security = $security;
     }
 
+    /**
+     * Retrieves all point telemetry records and returns them as a JSON response.
+     *
+     * @param EntityManagerInterface $entityManager The entity manager used to retrieve the records.
+     * @return JsonResponse The JSON response containing the telemetry data.
+     * @throws ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/point_telemetry', name: 'apiGetAllTelemetry', methods: ['GET'])]
     public function getAllTelemetry(EntityManagerInterface $entityManager): JsonResponse
     {
@@ -48,6 +57,13 @@ class PointTelemetryController extends AbstractController
         }
     }
 
+    /**
+     * Retrieves a specific point telemetry record by its ID and returns it as a JSON response.
+     *
+     * @param PointTelemetry $telemetry The point telemetry record to retrieve.
+     * @return JsonResponse The JSON response containing the telemetry data.
+     * @throws AccessDeniedException If the user is not authenticated.
+     */
     #[Route('/point_telemetry/{id}', name: 'apiGetTelemetry', methods: ['GET'])]
     public function getTelemetryById(PointTelemetry $telemetry): JsonResponse
     {
@@ -61,6 +77,16 @@ class PointTelemetryController extends AbstractController
         ], JsonResponse::HTTP_OK);
     }
 
+    /**
+     * Deletes a point telemetry record.
+     *
+     * @param PointTelemetry $telemetry The telemetry record to be deleted.
+     * @param EntityManagerInterface $entityManager The entity manager.
+     * @return JsonResponse The JSON response indicating the status of the deletion.
+     * @throws ConnectionException If there is a database connection error.
+     * @throws \Doctrine\DBAL\Exception If there is a database error.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/point_telemetry/{id}', name: 'apiDeleteTelemetry', methods: ['DELETE'])]
     public function deleteTelemetry(
         PointTelemetry $telemetry,
@@ -83,6 +109,18 @@ class PointTelemetryController extends AbstractController
         }
     }
 
+    /**
+     * Retrieves analytics data for point telemetry.
+     *
+     * This function retrieves analytics data for point telemetry by querying the database
+     * and calculating the total number of generated points, the number of visited points,
+     * and the visit rate for points generated by rule and without rule. The analytics data
+     * is returned in JSON format.
+     *
+     * @param EntityManagerInterface $entityManager The entity manager used to query the database.
+     * @return JsonResponse The JSON response containing the analytics data.
+     * @throws \Exception If there is an internal server error.
+     */
     #[Route('/point_telemetry_analytics', name: 'apiGetTelemetryAnalytics', methods: ['GET'])]
     public function getTelemetryAnalytics(EntityManagerInterface $entityManager): JsonResponse
     {
