+++ frontend/tochkarik/src/App.tsx
@@ -8,6 +8,21 @@ import AppRoutes from './routes/AppRoutes';
 import './App.css';
 
 
+/**
+ * Renders the main application component.
+ *
+ * This component initializes the authentication token from local storage and sets it as an
+ * HTTP-only cookie using the `setAuthToken` function. It also defines the `handleLogout`
+ * function to clear the authentication token from local storage and the `isAuthenticated`
+ * variable to check if a token is present in local storage.
+ *
+ * The component uses the `AuthProvider` and `Router` components from React Router to provide
+ * authentication context and routing functionality. The `AppRoutes` component is passed
+ * the `isAuthenticated`, `setAuthToken`, and `handleLogout` functions as props to handle
+ * authentication-related logic.
+ *
+ * @return {JSX.Element} The main application component.
+ */
 function App() {
     useEffect(() => {
         const token = localStorage.getItem('token');
