+++ frontend/tochkarik/src/components/Logout.tsx
@@ -5,6 +5,13 @@ interface LogoutProps {
     logoutHandler: () => void;
 }
 
+/**
+ * Renders a component that logs out the user and redirects them to the login page.
+ *
+ * @param {LogoutProps} props - The component props.
+ * @param {() => void} props.logoutHandler - The function to handle the logout.
+ * @return {React.ReactElement | null} Returns null as this component does not render anything.
+ */
 const Logout: React.FC<LogoutProps> = ({logoutHandler}) => {
     const navigate = useNavigate();
