+++ frontend/tochkarik/src/components/Map/GeneratedPoint/GeneratedPoint.tsx
@@ -9,7 +9,7 @@ import bookmarked from '../../../icons/bookmarked.svg';
 
 import "./GeneratedPoint.css";
 import '../Map.css';
-import {axiosInstance} from "../../../hooks/axiosConfig";
+import { axiosPrivateInstance } from "../../../services/authService";
 
 
 interface GeneratedPointProps {
@@ -27,6 +27,24 @@ interface GeneratedPointProps {
     onSave: (id: number | null) => void;
 }
 
+/**
+ * Renders a component for displaying a generated point with controls for starting a journey, creating a report, and bookmarking the point.
+ *
+ * @param {Object} props - The props object containing the following properties:
+ *   - pointId: The ID of the generated point.
+ *   - street: The street address of the generated point.
+ *   - pointTitle: The title of the generated point.
+ *   - isNew: A boolean indicating whether the generated point is new or not.
+ *   - hasReport: A boolean indicating whether the generated point has a report or not.
+ *   - onCancel: A function to handle cancellation of the generated point.
+ *   - onStartJourney: A function to handle starting a journey to the generated point.
+ *   - onCreateReport: A function to handle creating a report for the generated point.
+ *   - onEditReport: A function to handle editing a report for the generated point.
+ *   - coordinates: The coordinates of the generated point.
+ *   - timeOfGenerate: The time when the generated point was generated.
+ *   - onSave: A function to handle saving the generated point.
+ * @return {JSX.Element} The rendered component.
+ */
 const GeneratedPoint: React.FC<GeneratedPointProps> = ({
    pointId,
    street,
@@ -46,17 +64,28 @@ const GeneratedPoint: React.FC<GeneratedPointProps> = ({
     const [showAddPointDialog, setShowAddPointDialog] = useState<boolean>(false);
     const [localPointId, setLocalPointId] = useState<number | null>(pointId);
 
+    /**
+     * Handles adding a bookmark by showing the add point dialog and setting the bookmark status.
+     *
+     * @return {void} No return value.
+     */
     const handleAddBookmark = () => {
         setShowAddPointDialog(true);
         setIsBookmarked(true);
     };
 
+    /**
+     * Removes a bookmark from a point by sending a DELETE request to the server.
+     *
+     * @param {number | null} pointId - The ID of the point to remove the bookmark from.
+     * @return {void} This function does not return anything.
+     */
     const handleRemoveBookmark = (pointId: number | null) => {
         if (!pointId) {
             console.error("No pointId provided");
             return;
         }
-        axiosInstance.delete(`/points/${pointId}`)
+        axiosPrivateInstance.delete(`/points/${pointId}`)
             .then(() => {
                 setIsBookmarked(false);
                 onSave(null);
@@ -67,10 +96,21 @@ const GeneratedPoint: React.FC<GeneratedPointProps> = ({
             });
     }
 
+    /**
+     * Sets the state of `showAddPointDialog` to `false`, effectively closing the add point dialog.
+     *
+     * @return {void} This function does not return anything.
+     */
     const addButtonCancelHandler = () => {
         setShowAddPointDialog(false);
     };
 
+    /**
+     * A function that handles saving the point with the given ID.
+     *
+     * @param {number} id - The ID of the point to be saved.
+     * @return {void} No return value.
+     */
     const handleSave = (id: number) => {
         console.log('on save ' + id)
         setLocalPointId(id);
@@ -86,6 +126,11 @@ const GeneratedPoint: React.FC<GeneratedPointProps> = ({
         }
     }, [])
 
+    /**
+     * Opens a new tab in the user's browser with a Google Maps URL based on the provided coordinates.
+     *
+     * @return {void}
+     */
     const defaultOnStartJourney = () => {
         if (coordinates) {
             const [lat, lon] = coordinates;
