+++ frontend/tochkarik/src/components/Map/ManualLocation/ManualLocationInput.css
@@ -4,7 +4,6 @@
 }
 .manual-location-control {
     width: 100%;
-
     display: flex;
     justify-content: center;
 
@@ -18,6 +17,5 @@
 
 .search-suggestion-container li {
     font-family: "Helvetica Neue", Arial, Helvetica, sans-serif;
-
     margin-top: 10px;
 }
