+++ frontend/tochkarik/src/components/Map/Slider/RadiusSlider.tsx
@@ -6,6 +6,14 @@ interface RadiusSliderProps {
     handleRadiusChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
+/**
+ * Renders a radius slider component.
+ *
+ * @param {RadiusSliderProps} props - The props for the component.
+ * @param {number} props.radius - The current radius value.
+ * @param {function} props.handleRadiusChange - The function to handle radius change.
+ * @return {ReactElement} The rendered radius slider component.
+ */
 const RadiusSlider: React.FC<RadiusSliderProps> = ({ radius, handleRadiusChange }) => (
     <div className="slider-container">
         <input
