--- frontend/tochkarik/src/hooks/api.token.ts
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-const setAuthToken = (token: string | null): void => {
-    if (token) {
-        axios.defaults.headers.common["Authorization"] = `Bearer ${token}`;
-    } else {
-        delete axios.defaults.headers.common["Authorization"];
-    }
-}
-
-export default setAuthToken;
