+++ frontend/tochkarik/src/main.tsx
@@ -3,22 +3,9 @@ import ReactDOM from 'react-dom/client'
 import App from './App'
 import './index.css'
 
-
-
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
         <App />
     </React.StrictMode>,
 )
-//
-// if ('serviceWorker' in navigator) {
-//     window.addEventListener('load', () => {
-//         navigator.serviceWorker.register('/service-worker.js')
-//             .then(registration => {
-//                 console.log('ServiceWorker registration successful with scope: ', registration.scope);
-//             }, err => {
-//                 console.log('ServiceWorker registration failed: ', err);
-//             });
-//     });
-// }
