+++ frontend/tochkarik/src/pages/Bookmarks/Bookmarks.tsx
@@ -1,6 +1,5 @@
 import {FC, useEffect, useState} from 'react';
-import {setAuthToken} from '../../hooks/axiosConfig';
-import {axiosPrivateInstance} from '../../services/authService'
+import {axiosPrivateInstance, setAuthToken} from '../../services/authService'
 
 import GeneratedPoint from "../../components/Map/GeneratedPoint/GeneratedPoint";
 import MapComponent from "../../components/Map/MapComponent";
@@ -17,12 +16,23 @@ interface Bookmark {
     description: string;
 }
 
+/**
+ * Renders the Bookmarks component with dynamic content based on the fetched bookmarks.
+ * Handles fetching bookmarks, canceling the selection, and deleting a specific bookmark.
+ *
+ * @return {JSX.Element} The Bookmarks component JSX.
+ */
 const Bookmarks: FC = () => {
     const [bookmarks, setBookmarks] = useState<Bookmark[]>([]);
     const [loading, setLoading] = useState<boolean>(true);
     const [error, setError] = useState<string | null>(null);
     const [selectedBookmark, setSelectedBookmark] = useState<Bookmark | null>(null);
 
+    /**
+     * Fetches the bookmarks from the server and updates the state.
+     *
+     * @return {Promise<void>} A Promise that resolves when the bookmarks are fetched and the state is updated.
+     */
     const fetchBookmarks = () => {
         const token = localStorage.getItem('token');
 
@@ -40,6 +50,11 @@ const Bookmarks: FC = () => {
         });
     };
 
+    /**
+     * Resets the selected bookmark to null and fetches the bookmarks from the server.
+     *
+     * @return {void} This function does not return anything.
+     */
     const handleCancel = () => {
         setSelectedBookmark(null);
         fetchBookmarks();
@@ -49,6 +64,12 @@ const Bookmarks: FC = () => {
         fetchBookmarks();
     }, []);
 
+    /**
+     * Handles the deletion of a point by filtering out the bookmark with the specified ID and resetting the selected bookmark.
+     *
+     * @param {number | null} pointId - The ID of the point to be deleted.
+     * @return {void} This function does not return anything.
+     */
     const handlePointDeletion = (pointId: number | null) => {
         if (pointId) {
             setBookmarks(bookmarks.filter(bookmark => bookmark.id !== pointId));
