+++ frontend/tochkarik/src/pages/UserPosts/UserPosts.tsx
@@ -6,6 +6,11 @@ import PostPreview from "../../components/Post/PostPreview"
 
 
 
+/**
+ * Renders the user posts component.
+ *
+ * @return {ReactElement} The rendered user posts component.
+ */
 const UserPosts: FC = () => {
     const navigate = useNavigate();
     const redirect = (path: string) => {
