+++ frontend/tochkarik/src/routes/AuthenticatedRoutes.tsx
@@ -6,7 +6,7 @@ import Header from '../components/Header/Header';
 import UserProfile from '../pages/UserProfile/UserProfile'
 import UserPosts from '../pages/UserPosts/UserPosts'
 import Post from '../components/Post/Post'
-import NewHomePage from '../pages/HomePage/NewHomePage';
+import HomePage from '../pages/HomePage/HomePage';
 import Bookmarks from '../pages/Bookmarks/Bookmarks'
 import AdminDashboard from "../pages/AdminDashboard/AdminDashboard";
 
@@ -14,14 +14,18 @@ interface AuthenticatedRoutesProps {
     logoutHandler: () => void;
 }
 
-
+/**
+ * Renders the authenticated routes for the application.
+ *
+ * @param {AuthenticatedRoutesProps} props - The props object containing the logoutHandler.
+ * @return {JSX.Element} The JSX element representing the authenticated routes.
+ */
 const AuthenticatedRoutes: FC<AuthenticatedRoutesProps> = ({ logoutHandler }) => {
     return (
         <>
             <Header/>
-
             <Routes>
-                <Route path="/" element={<NewHomePage />} />
+                <Route path="/" element={<HomePage/>} />
                 <Route path="/profile" element={<UserProfile userId={1} logoutHandler={logoutHandler} />} />
                 <Route path="/user_posts" element={ <UserPosts /> } />
                 <Route path="/post" element={ <Post /> } />
