+++ frontend/tochkarik/src/routes/UnauthenticatedRoutes.tsx
@@ -2,7 +2,7 @@ import React, {FC} from 'react';
 import {Navigate, Route, Routes} from 'react-router-dom';
 
 import LoginPage from "../pages/LoginPage/LoginPage";
-import NewHomePage from "../pages/HomePage/NewHomePage";
+import HomePage from "../pages/HomePage/HomePage";
 import RegistrationPage from "../pages/RegisterPage/RegistrationPage";
 
 import Header from "../components/Header/Header";
@@ -11,12 +11,18 @@ interface UnauthenticatedRoutesProps {
     setAuthToken: (token: string | null) => void;
 }
 
+/**
+ * Renders the routes for unauthenticated users.
+ *
+ * @param {FC<UnauthenticatedRoutesProps>} setAuthToken - Function to set authentication token.
+ * @return {ReactElement} The rendered routes for unauthenticated users.
+ */
 const UnauthenticatedRoutes: FC<UnauthenticatedRoutesProps> = ({ setAuthToken }) => {
     return (
         <>
             <Header user_login={false} />
             <Routes>
-                <Route path="/" element={<NewHomePage />} />
+                <Route path="/" element={<HomePage/>} />
                 <Route path="/login" element={<LoginPage setAuthToken={setAuthToken} />} />
                 <Route path="/reg" element={<RegistrationPage setAuthToken={setAuthToken} />} />
                 <Route path="*" element={<Navigate to="/login" />} />
