+++ frontend/tochkarik/src/services/authService.tsx
@@ -11,6 +11,12 @@ const axiosPrivateInstance: AxiosInstance = axios.create({
 });
 
 
+/**
+ * Sets the authorization token in the axiosPrivateInstance's default headers.
+ *
+ * @param {string | null} token - The token to be set as the authorization header. If null, the authorization header is deleted.
+ * @return {void} This function does not return anything.
+ */
 export const setAuthToken = (token: string | null) => {
     if (token) {
         axiosPrivateInstance.defaults.headers.common["Authorization"] = `Bearer ${token}`;
@@ -19,6 +25,12 @@ export const setAuthToken = (token: string | null) => {
     }
 };
 
+/**
+ * Refreshes the authentication token.
+ *
+ * @return {Promise<string | null>} A promise that resolves to the new token if successful, or null if the refresh failed.
+ * @throws {Error} If no refresh token is available.
+ */
 export const refreshToken = async (): Promise<string | null> => {
     const refreshToken = localStorage.getItem('refresh_token');
     if (!refreshToken) throw new Error("No refresh token available");
@@ -38,6 +50,11 @@ export const refreshToken = async (): Promise<string | null> => {
     })
 };
 
+/**
+ * Clears the authentication tokens from the local storage and sets the authorization token in the axiosPrivateInstance's default headers to null.
+ *
+ * @return {void} This function does not return anything.
+ */
 export const clearAuthTokens = () => {
     localStorage.removeItem('token');
     localStorage.removeItem('refresh_token');
@@ -45,6 +62,11 @@ export const clearAuthTokens = () => {
     window.location.href = '/';
 };
 
+/**
+ * Handles the logout process.
+ *
+ * @return {void} This function does not return anything.
+ */
 const logoutHandler = () => {
     if (window.location.pathname !== '/login') {
         clearAuthTokens();
@@ -52,7 +74,13 @@ const logoutHandler = () => {
     }
 };
 
-// Interceptor for the private API instance
+/**
+ * Interceptor for handling responses in axiosPrivateInstance.
+ *
+ * @param {Object} response - The response object.
+ * @param {Object} error - The error object.
+ * @return {Promise} A Promise that resolves the response or rejects the error.
+ */
 axiosPrivateInstance.interceptors.response.use(
     response => response,
     async (error) => {
